refactor: added auth controller
