+++ server/index.js
@@ -14,6 +14,7 @@ require('./services/passport');
 
 const clientRoutes = require('./routes/clientRoutes');
 const assetRoutes = require('./routes/assetRoutes');
+const authRoutes = require('./routes/authRoutes');
 
 // connect to the database
 const connDb = mongoose.connect(keys.mongoURI).then(
@@ -44,7 +45,7 @@ app.use(passport.session());
 app.use(morgan('dev'));
 
 // import routes
-require('./routes/authRoutes')(app);
+app.use('/api/whd/auth', authRoutes);
 app.use('/api/whd/clients', clientRoutes);
 app.use('/api/whd/assets', assetRoutes);
